requirements.txt
# Core ML libraries
torch>=2.0.0
torchaudio>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
timm>=1.0.0
datasets>=2.10.0
librosa>=0.10.0
scipy>=1.10.0
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.3.0
soundfile>=0.12.0

# Metrics and evaluation
jiwer>=3.0.0
sacrebleu>=2.3.0
rouge-score>=0.1.2
nltk>=3.8.0
evaluate>=0.4.0

# API and web
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
python-multipart>=0.0.6
requests>=2.31.0

# Data processing
seaborn>=0.12.0
plotly>=5.14.0
kaleido>=0.2.1

# Environment and utilities
python-dotenv>=1.0.0
pathlib
argparse
logging
tqdm>=4.65.0

# Speech models specific
accelerate>=0.20.0
safetensors>=0.3.0

# SeamlessM4T dependencies
tiktoken>=0.5.0
blobfile>=2.1.0
sentencepiece>=0.1.99

# Optional but recommended for performance
# optimum>=1.9.0
# onnxruntime>=1.15.0
protobuf



# Added from the code block
speechbrain>=0.5.15 